pyproject.toml
[tool.poetry]
name = "api.trampoja.com"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8.11"
Django = "3.2.5"
djangorestframework = "3.12.4"
httpie = "2.4.0"
Pillow = "8.3.0"
requests = "2.25.1"
django-cors-headers = "3.7.0"
gunicorn = "20.1.0"
celery = "5.1.2"
redis = "3.5.3"
psycopg2-binary = "2.8.6"
psycopg2 = "2.8.6"
python-dotenv = "0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
